backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
sentence-transformers
faiss-cpu
pandas
openpyxl
pydantic
python-dotenv